refactor(vscode): extract editor actions from provider
